Add license file
From http://unbuffered.info/danbooruup/:
"As the code for the native component is partly swiped from the Mozilla
source, the extension is licensed under the MPL version 1.1."

The native component is gone now, but of course the license remains.
There aren't any Exhibit A notices in the remaining source code, and I
elect not to add them.